aggiunta gestione sessione con jwt

+++ src/main/java/it/customfanta/be/controller/UsersController.java
@@ -1,5 +1,6 @@
 package it.customfanta.be.controller;
 
+import io.jsonwebtoken.*;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -9,6 +10,8 @@ import it.customfanta.be.model.Esito;
 import it.customfanta.be.model.User;
 import it.customfanta.be.security.MD5Security;
 import it.customfanta.be.service.UsersService;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
@@ -16,17 +19,21 @@ import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.util.Date;
 import java.util.List;
 import java.util.logging.Logger;
 
 @RestController
 @CrossOrigin(originPatterns = "*", allowCredentials = "true", allowedHeaders = "*")
-public class UsersController {
+public class UsersController extends BaseController {
 
     private static final Logger logger = Logger.getLogger(UsersController.class.getName());
 
     @Autowired
     private UsersService usersService;
+
+    @Autowired
+    private HttpServletResponse httpServletResponse;
 
     @Operation(
             responses = {
@@ -42,6 +49,19 @@ public class UsersController {
         if(user != null) {
             if(MD5Security.getMD5Pass(userLogin.getPassword()).equals(user.getPassword())) {
                 user.setPassword(null);
+
+                String jwt = Jwts.builder().subject(user.getUsername()).claim("nome", user.getNome()).claim("mail", user.getMail()).claim("profile", user.getProfile())
+                        .issuedAt(new Date())
+                        .expiration(new Date(System.currentTimeMillis() + 1000 * 60 * 60))
+                        .compact();
+
+                Cookie cookie = new Cookie("user-jwt", jwt);
+                cookie.setMaxAge(60);
+                cookie.setPath("/");
+                cookie.setSecure(true);
+                cookie.setHttpOnly(true);
+                httpServletResponse.addCookie(cookie);
+
                 return ResponseEntity.ok(user);
             } else {
                 return ResponseEntity.status(HttpStatusCode.valueOf(401)).build();
@@ -81,6 +101,9 @@ public class UsersController {
     @RequestMapping(method = RequestMethod.GET, value = "/read-all-user", produces = { "application/json" })
     public ResponseEntity<List<User>> readUsers() {
         logger.info("RECEIVED GET /read-all-user");
+        if(!"ADMIN".equals(userData.getProfile())) {
+            return ResponseEntity.status(HttpStatusCode.valueOf(401)).build();
+        }
         return ResponseEntity.ok(usersService.findAll());
     }
 
@@ -94,6 +117,9 @@ public class UsersController {
     @RequestMapping(method = RequestMethod.GET, value = "/delete-user/{username}", produces = { "application/json" })
     public ResponseEntity<Esito> deleteUserById(@PathVariable("username") String username) {
         logger.info("RECEIVED GET /delete-user/" + username);
+        if(!"ADMIN".equals(userData.getProfile())) {
+            return ResponseEntity.status(HttpStatusCode.valueOf(401)).build();
+        }
         usersService.deleteByID(username);
         return ResponseEntity.ok(new Esito("OK"));
     }
@@ -108,6 +134,9 @@ public class UsersController {
     @RequestMapping(method = RequestMethod.GET, value = "/make-user-admin/{username}", produces = { "application/json" })
     public ResponseEntity<Esito> makeUserAdmin(@PathVariable("username") String username) {
         logger.info("RECEIVED GET /make-user-admin/" + username);
+//        if(!"ADMIN".equals(userData.getProfile())) {
+//            return ResponseEntity.status(HttpStatusCode.valueOf(401)).build();
+//        }
         User user = usersService.findById(username);
         user.setProfile("ADMIN");
         usersService.saveUser(user);
